refactor (ConfirmationDIalog): Change the location of the dialog and deletion operation to only one place using context
