+++ public/index.html
@@ -26,7 +26,7 @@
       href="https://fonts.googleapis.com/css?family=Roboto"
       rel="stylesheet"
     />
-    <title>React App</title>
+    <title>Conference App</title>
   </head>
   <body>
     <noscript>
