+++ src/providers/ModalForm/ModalForm.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { Modal, withStyles } from "@material-ui/core";
 import { BookingForm } from "components/Modals/CreateMeeting/BookingForm";
+import { ConfirmationDialog } from "components/Modals/DeleteBooking/ConfirmationDialog";
 import { storageService } from "services";
 
 const styles = theme => ({
@@ -10,12 +11,18 @@ const styles = theme => ({
 const ModalFormContext = React.createContext({
   handleOnClickCreateMeeting: () => {},
   handleOnClickEditMeeting: () => {},
-  handleOnCloseModal: () => {}
+  handleOnCloseModal: () => {},
+  handleDeleteMeeting: () => {},
+  handleCloseDialog: () => {}
 });
 
 class ModalFormProviderComponent extends React.Component {
   state = {
-    isModalOpen: false
+    isModalOpen: false,
+    isDialogOpen: false,
+    isDelete: false,
+    isDeleteLoading: false,
+    shouldUpdate: false
   };
 
   handleClickCreateBooking = booking => {
@@ -31,6 +38,7 @@ class ModalFormProviderComponent extends React.Component {
     const { roomName } = booking;
     this.setState({
       isModalOpen: true,
+      isDelete: false,
       room: roomName,
       bookingClicked: false,
       quickAppointment: true,
@@ -42,6 +50,7 @@ class ModalFormProviderComponent extends React.Component {
   handleCreationFromButton = () => {
     this.setState({
       isModalOpen: true,
+      isDelete: false,
       room: null,
       bookingClicked: false,
       quickAppointment: false,
@@ -67,18 +76,42 @@ class ModalFormProviderComponent extends React.Component {
     }
   };
 
+  handleDeleteBooking = booking => {
+    this.setState(prev => ({
+      shouldUpdate: !prev.shouldUpdate,
+      isDelete: true,
+      isDialogOpen: true,
+      bookingFormated: booking
+    }));
+  };
+
+  handleDeleteLoading = () => {
+    this.setState({ isDeleteLoading: true });
+  };
+
+  handleAfterDelete = () => {
+    this.setState({ isDeleteLoading: false });
+  };
+
   handleModalClose = () => {
     this.setState({ isModalOpen: false });
+  };
+
+  handleDialogClose = () => {
+    this.setState({ isDialogOpen: false, isDelete: false });
   };
 
   render() {
     const {
       isModalOpen,
+      isDialogOpen,
+      isDeleteLoading,
       bookingClicked,
       bookingClickedObj,
       quickAppointment,
       roomName,
-      roomId
+      roomId,
+      bookingFormated
     } = this.state;
 
     const {
@@ -92,25 +125,40 @@ class ModalFormProviderComponent extends React.Component {
         value={{
           handleOnClickCreateMeeting: this.handleClickCreateBooking,
           handleOnClickEditMeeting: this.handleClickEditBooking,
-          handleOnCloseModal: this.handleModalClose
+          handleOnCloseModal: this.handleModalClose,
+          handleDeleteMeeting: this.handleDeleteBooking,
+          handleCloseDialog: this.handleDialogClose
         }}
       >
-        <Modal
-          className={modal}
-          open={isModalOpen}
-          disableAutoFocus={true}
-          onClose={this.handleModalClose}
-        >
-          <BookingForm
-            roomName={roomName}
-            roomId={roomId}
-            isBookingEdition={bookingClicked}
-            bookingForEdition={bookingClickedObj}
-            quickAppointment={quickAppointment}
-            onModalClose={this.handleModalClose}
+        {this.state.isDelete ? (
+          <ConfirmationDialog
+            bookingInfo={bookingFormated}
+            isOpen={isDialogOpen}
+            isLoading={isDeleteLoading}
+            onCancel={this.handleDialogClose}
             onBookingsDataChange={onBookingsDataChange}
+            update={this.state.shouldUpdate}
+            onDeleteLoading={this.handleDeleteLoading}
+            onAfterDelete={this.handleAfterDelete}
           />
-        </Modal>
+        ) : (
+          <Modal
+            className={modal}
+            open={isModalOpen}
+            disableAutoFocus={true}
+            onClose={this.handleModalClose}
+          >
+            <BookingForm
+              roomName={roomName}
+              roomId={roomId}
+              isBookingEdition={bookingClicked}
+              bookingForEdition={bookingClickedObj}
+              quickAppointment={quickAppointment}
+              onModalClose={this.handleModalClose}
+              onBookingsDataChange={onBookingsDataChange}
+            />
+          </Modal>
+        )}
         {children}
       </ModalFormContext.Provider>
     );
